compatibility ratios now adjust according to changes in selected interests
